Updates caching tests
